requirements.txt
aeon==0.5.0
captum==0.6.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
joypy==0.2.6
lightning==2.0.2
matplotlib==3.7.1
more-itertools==9.1.0
numpy==1.21.4
omegaconf==2.3.0
optuna==2.10.1
pandas==2.0.1
petastorm==0.12.1
pre-commit==3.2.2
pyarrow==10.0.1
pyrootutils==1.0.4
pyspark==3.4.0
pytest==7.3.1
PyYAML==6.0
rich==13.3.4
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
tensorboard==2.12.2
torch==1.13.1
torchmetrics==0.11.4
tqdm==4.65.0
